#COBUMMFO-8124# Коментар: MERGE:COBUMMFO-8124 Ростислав Кузьмович

+++ web/mmfo/barsroot/App_Code/Areas/BpkW4/Controllers/СheckdkboController.cs
@@ -62,7 +62,7 @@ namespace BarsWeb.Areas.BpkW4.Controllers
         }
 
         [HttpGet]
-        public ActionResult QuestItemsDisplay(int dealId, string customerRnk)
+        public ActionResult QuestItemsDisplay(decimal dealId, string customerRnk)
         {
             W4_DKBO_WEB_FilterParams fp = new W4_DKBO_WEB_FilterParams() { DEAL_ID = dealId, CUSTOMER_ID = customerRnk};
             return RedirectToAction("QuestItemsShow", fp);
